refactor(tui): split hud.rs catch-all into 5 painter modules (#575)
Closes #575 — the G-audit Divergent-Change finding. hud.rs (1224 lines)
held FIVE unrelated ratatui painters, changing for five unrelated reasons.
Split into cohesive modules under widgets/ (pure code moves, no logic
change):
- theme_picker.rs (theme_swatch + paint_theme_picker)
- footer.rs (FooterStats, source_warning_message, paint_footer, status
  segments/tiers, build_status_*)
- wall_board.rs (BOARD_W, board geometry, paint_wall_display + star_hit_rect
  — kept together: the paint↔hit-rect pairing is geometry-coupled and
  sharp-edge-pinned)
- version_popup.rs (REPO_URL, VERSION_POPUP_URL, the popup + url-rect)
- elevator.rs (paint_elevator_indicator)

hud.rs deleted; widgets/mod.rs re-exports repathed to the 5 modules (every
external caller reaches these via the widgets:: re-exports, so no call-site
code changed). Each moved item keeps its exact visibility (unreachable_pub
forced no demotions). Tests co-located with their painters (16 moved, split
3/2/2/8/1). Docs swept tree-wide to stale-grep==0 (tui/CLAUDE.md
hud::star_hit_rect → wall_board::star_hit_rect, + every other hud location
mention across both crates' guides).

Behavior-preserving: 2310/2310 (test count parity), zero .snap/media churn,
clippy -D warnings clean.
